Make #gt jump to the `#`-nth elscreen tab
This is a little confusing, because on Vim tabs are 1-indexed, while
elscreen tabs are 0-indexed. However, typing 0gt on `evil-mode` passed
the `count` argument as `nil`, rather than `0`. This makes it hard (with
my lack of Emacs Lisp knowledge) to make 0gt work properly and make the
feature use 0-indexing on emacs - tabs are numbered by default, so this
would obviously be better.

Mimicking the Vim behaviour is not so bad, though.